Schematic and photos
The breadboard circuit is composed of 2 IDT7132 dual port 2K x 8 RAM chips replacing the Pinball machine U25RAM.
An Atmega1284 connects to the A side ports of the RAM.
The Atmega1284 runs a command line program to receive requests over serial UART from the ESP32.
The 6802 processor of the Pinball machine has full access to the primary RAM to avoid memory access conflicts with the Atmega1284.

A future revision may use a clock stretching wait state on the 6802 to use the Busy signal from the RAM to eliminate any chance of memory contention.

The critical memory addresses in the RAM is the area used for the program stack to call and return from subroutines and interrupts.

During game play the ESP32 is limited to reading memory from the addresses related to game scores, game status, and player number.

As of 2023-04-11 the Busy signals from RAM cause an interrupt to the Atmega1284 to count the occurrence. A command was added to the code to query the counters. A later revision of the code will have recovery options if there is a bus conflict detected.